seperate functions from tasksTable and bind them to the class Relates #252
